Improving exception handling to make it easier and isolate exception concerns from everywhere else. Plus, definining the first draft of the DDD architecture, yay!
