+++ config/services.yaml
@@ -15,7 +15,6 @@ services:
     # this creates a service per class whose id is the fully-qualified class name
     App\:
         resource: '../src/*'
-        exclude: '../src/{Entity,Migrations,Tests}'
 
     # controllers are imported separately to make sure services can be injected
     # as action arguments even if you don't extend any base controller class
@@ -25,7 +24,7 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
-    App\EventListener\ExceptionListener:
+    App\Infrastructure\EventListener\Symfony\ExceptionListener:
         arguments: ['%env(APP_ENV)%']
         tags:
             - { name: kernel.event_listener, event: kernel.exception }
